fixed gmx typo, fixed test names, remove console logs
